# api/openapi.yaml
openapi: 3.1.0
info:
  title: Docket API
  version: 1.0.0
  description: |
    The API behind Docket, an admissions document verification tool. Staff import applications from a CSV and upload scanned documents, the service reads each document and compares it with the application, and verifiers decide the applications that are flagged. It is called by one first-party React frontend. Every example uses synthetic data.
    Serves: US-00-001, US-00-002, US-00-003, US-00-004, US-00-005, US-00-006, US-00-007, US-00-008, US-00-009, US-00-011, US-02-001. Style: docs/api/README.md.
    This spec is a design: no route except the two health routes exists yet.
  contact:
    name: backend lead rotation
  x-api-style: REST with JSON bodies, plus multipart uploads and one CSV download; one service, no GraphQL.
  x-versioning: >-
    The path carries no major version: the API has one client, built and deployed with it, so nothing installed needs to keep working.
    X-API-Version is accepted on every operation for dated behaviour changes and the server default is the newest.
    A second client would reopen this and add /v1.
  x-authentication: >-
    A server-side cookie session (docket_session, ADR-0006) on every operation except those marked security: [] (sign-in and health).
    State-changing requests also need an anti-forgery token. A role that is not allowed gets 403 and an id that does not exist gets 404.
  x-conventions:
    - rule: "No /v1 in the path."
...
servers:
  - url: /api
...
security:
  - cookieAuth: []
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      summary: List applications, optionally only the review queue
      description: >-
        Newest change first. The review queue is filter[status]=needs_review with filter[rejected]=false. Erased applications are never listed.
      tags: [applications]
      x-story-ids: [US-00-006, US-00-008]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: filter[status]
          in: query
          description: Only applications with this status.
          schema: { type: string, enum: [verified, needs_review, missing_documents] }
        - name: filter[rejected]
          in: query
          description: With needs_review, false is the queue and true is the rejected ones.
          schema: { type: boolean }
      responses:
        '200':
          description: One page
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Application' }
                  page: { $ref: '#/components/schemas/Page' }
              example:
                data:
                  - id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
                    application_ref: SYN-APP-004
                    full_name: Latha Sharma
                    father_name: Salim Sharma
                    date_of_birth: 2006-12-10
                    board: Karnataka State Board
                    roll_number: SYN0945957
                    marks: { English: 59, Hindi: 77, Mathematics: 87, Science: 74, Social Science: 56 }
                    category: General
                    status: needs_review
                    rejected: false
                    created_at: 2026-10-05T09:00:00Z
                    updated_at: 2026-10-05T09:12:41Z
                page: { next_cursor: eyJ1IjoiMjAyNiJ9, has_more: false }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
  /applications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getApplication
      summary: Get one application with its documents and per-field comparison
      description: Shows each document beside the application value, field by field, and which fields sent the application to review.
      tags: [applications]
      x-story-ids: [US-00-006]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The application
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            ETag: { schema: { type: string }, description: "Version of the application (its updated_at). Send it as If-Match on a decision." }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /applications/{id}/documents:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listDocuments
      summary: List an application's documents
      tags: [documents]
      x-story-ids: [US-00-002, US-00-003]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: One page, newest first
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Document' }
                  page: { $ref: '#/components/schemas/Page' }
              example:
                data:
                  - id: 0192b1c3-1a2b-7c4e-9a1b-2c3d4e5f6a7c
                    application_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
                    detected_type: 10th_marksheet
                    status: read
                    failure_reason: null
                    is_current: true
                    created_at: 2026-10-05T09:05:00Z
                page: { next_cursor: null, has_more: false }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: uploadDocument
      summary: Upload one scanned document for an application
      description: >-
        Staff role only. JPEG, PNG or PDF up to 8 MiB, checked by declared type and by the first bytes of the file.
        The file is stored and read in the background. Identical bytes already stored for this application return 200 with the existing document.
      tags: [documents]
      x-story-ids: [US-00-002]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The scan or photo.
            encoding:
              file: { contentType: "image/jpeg, image/png, application/pdf" }
      responses:
        '200':
          description: The same bytes were already stored for this application
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '202':
          description: Stored; reading has been queued
          headers:
            Location: { schema: { type: string, format: uri } }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /documents/{id}/image:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getDocumentImage
      summary: Get the stored image of a document
      description: Always a re-encoded JPEG or PNG with a fixed content type, Cache-Control private no-store and X-Content-Type-Options nosniff. A PDF is never served.
      tags: [documents]
      x-story-ids: [US-00-006]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The image bytes
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { schema: { type: string, example: "private, no-store" } }
            X-Content-Type-Options: { schema: { type: string, example: nosniff } }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
            image/png:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: docket_session
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      description: Date version, for example 2026-10-05. Missing means the server default, the newest.
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client UUID made once per logical operation and reused on every retry of it. The same key and body replays the first result for 24 hours, a different body under the same key is refused with 409 idempotency_conflict (stored in idempotency_keys).
      schema: { type: string, format: uuid }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Anti-forgery token for a state-changing request with a cookie session. Copy the docket_csrf cookie value into this header (ADR-0012). The server also refuses an Origin that is not allowed.
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Page size, at most 100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  headers:
    XRequestId:
      schema: { type: string }
      description: Echoed or generated request id; quote it in support tickets.
  schemas:
    Error:
      type: object
      required: [error]
      description: >-
        The one envelope for every non-2xx response (app/core/errors.py:53). The code is stable, the message is for people.
      x-error-codes:
        - { code: bad_request, status: 400, meaning: "Malformed request." }
        - { code: unauthorized, status: 401, meaning: "No session, an expired one, a revoked one or a wrong password." }
        - { code: forbidden, status: 403, meaning: "Signed in, but this role may not do this." }
        - { code: not_found, status: 404, meaning: "No such record, or the route does not exist." }
        - { code: conflict, status: 409, meaning: "The request conflicts with the current state." }
        - { code: idempotency_conflict, status: 409, meaning: "Idempotency-Key reused with a different body." }
        - { code: payload_too_large, status: 413, meaning: "The file is over the size cap." }
        - { code: unsupported_media_type, status: 415, meaning: "The declared type is not allowed or does not match the file." }
        - { code: validation_error, status: 422, meaning: "The request failed validation, or the content breaks a rule (details names it)." }
        - { code: rate_limited, status: 429, meaning: "Too many attempts; wait Retry-After seconds." }
        - { code: internal, status: 500, meaning: "Unexpected failure; quote request_id." }
        - { code: service_unavailable, status: 503, meaning: "A dependency the request needs is not reachable." }
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string, example: validation_error }
            message: { type: string, example: request failed validation }
            details:
              type: object
              additionalProperties: true
              description: Structured detail. For validation_error it holds an errors list with loc, msg and type, never the submitted value.
              example: { errors: [{ loc: [body, reason], msg: "Field required", type: missing }] }
            request_id:
              type: [string, 'null']
              example: 4f1c0a8e3b7d4a0c9b6e2d1f5a7c8e90
    Page:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
      example: { next_cursor: eyJ1IjoiMjAyNiJ9, has_more: false }
...
    Application:
      type: object
      required: [id, application_ref, full_name, father_name, date_of_birth, board, roll_number, marks, category, status, rejected, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        application_ref: { type: string, maxLength: 40 }
        full_name: { type: string }
        father_name: { type: string }
        date_of_birth: { type: string, format: date }
        board: { type: string }
        roll_number: { type: string }
        marks: { type: object, additionalProperties: { type: integer, minimum: 0, maximum: 100 } }
        category: { type: string }
        status: { type: string, enum: [verified, needs_review, missing_documents] }
        rejected: { type: boolean, description: "True when a verifier rejected it. It stays needs_review but leaves the queue." }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        application_ref: SYN-APP-004
        full_name: Latha Sharma
        father_name: Salim Sharma
        date_of_birth: 2006-12-10
        board: Karnataka State Board
        roll_number: SYN0945957
        marks: { English: 59, Hindi: 77, Mathematics: 87, Science: 74, Social Science: 56 }
        category: General
        status: needs_review
        rejected: false
        created_at: 2026-10-05T09:00:00Z
        updated_at: 2026-10-05T09:12:41Z
    ExtractedField:
      type: object
      required: [id, field_name, value, needs_review]
      properties:
        id: { type: integer, format: int64 }
        field_name: { type: string, enum: [name, father_name, dob, board, roll_number, marks, document_number] }
        subject: { type: [string, 'null'], description: "The subject for a marks row." }
        value: { type: string }
        application_value: { type: [string, 'null'], description: "The value on the application this field is compared to, so the screen can show both side by side. Null when the application holds nothing to compare, for example a document number." }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        box:
          type: [array, 'null']
          description: Four numbers, x, y, width and height in pixels of the stored image.
          items: { type: number }
        match_result: { type: [string, 'null'], enum: [match, mismatch, skipped, null] }
        needs_review: { type: boolean }
        review_reason: { type: [string, 'null'], enum: [low_confidence, format_invalid, mismatch, not_extracted, null] }
      example:
        id: 4012
        field_name: marks
        subject: Mathematics
        value: "87"
        confidence: 0.9971
        box: [320, 372, 40, 22]
        match_result: match
        needs_review: false
        review_reason: null
    Document:
      type: object
      required: [id, application_id, status, is_current, created_at]
      properties:
        id: { type: string, format: uuid }
        application_id: { type: string, format: uuid }
        detected_type: { type: [string, 'null'], enum: [10th_marksheet, 12th_marksheet, id_proof, transfer_certificate, unknown, null] }
        status: { type: string, enum: [uploaded, processing, read, failed] }
        failure_reason: { type: [string, 'null'], description: "A short reason code, never document content." }
        is_current: { type: boolean }
        created_at: { type: string, format: date-time }
        fields:
          type: array
          description: Present on the application detail, empty until the document is read.
          items: { $ref: '#/components/schemas/ExtractedField' }
      example:
        id: 0192b1c3-1a2b-7c4e-9a1b-2c3d4e5f6a7c
        application_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        detected_type: 10th_marksheet
        status: read
        failure_reason: null
        is_current: true
        created_at: 2026-10-05T09:05:00Z
    ApplicationDetail:
      allOf:
        - $ref: '#/components/schemas/Application'
        - type: object
          required: [documents]
          properties:
            documents:
              type: array
              items: { $ref: '#/components/schemas/Document' }
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        application_ref: SYN-APP-004
        full_name: Latha Sharma
        father_name: Salim Sharma
        date_of_birth: 2006-12-10
        board: Karnataka State Board
        roll_number: SYN0945957
        marks: { English: 59, Hindi: 77, Mathematics: 87, Science: 74, Social Science: 56 }
        category: General
        status: needs_review
        rejected: false
        created_at: 2026-10-05T09:00:00Z
        updated_at: 2026-10-05T09:12:41Z
        documents:
          - id: 0192b1c3-1a2b-7c4e-9a1b-2c3d4e5f6a7c
            application_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
            detected_type: 10th_marksheet
            status: read
            failure_reason: null
            is_current: true
            created_at: 2026-10-05T09:05:00Z
            fields:
              - { id: 4012, field_name: marks, subject: Mathematics, value: "87", confidence: 0.9971, box: [320, 372, 40, 22], match_result: match, needs_review: false, review_reason: null }
...
  responses:
    Unauthorized:
      description: No session, an invalid or expired one, or a wrong password
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Signed in, but this role may not do this
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such record
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    IdempotencyConflict:
      description: Conflict. The Idempotency-Key was reused with a different body (code idempotency_conflict), or the record is not in the state this call needs, for example an application that changed since the If-Match version or a document that is not failed (code conflict)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    PayloadTooLarge:
      description: The file is over the size cap
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    UnsupportedMediaType:
      description: The type is not allowed or does not match the file
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationFailed:
      description: The request failed validation or breaks a rule
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    InternalError:
      description: Unexpected failure; request_id is in the body
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...